extracted handling of ListRecords to a separate class
