+++ oai-pmh-handler/src/main/java/no/sikt/nva/oai/pmh/handler/AbstractOaiPmhMethodHandler.java
@@ -1,0 +1,24 @@
+package no.sikt.nva.oai.pmh.handler;
+
+import jakarta.xml.bind.JAXBElement;
+import java.time.ZonedDateTime;
+import java.util.GregorianCalendar;
+import javax.xml.datatype.DatatypeFactory;
+import org.openarchives.oai.pmh.v2.OAIPMHtype;
+import org.openarchives.oai.pmh.v2.ObjectFactory;
+
+public class AbstractOaiPmhMethodHandler {
+  private static final ObjectFactory objectFactory = new ObjectFactory();
+
+  protected JAXBElement<OAIPMHtype> baseResponse() {
+    var responseDate =
+        DatatypeFactory.newDefaultInstance()
+            .newXMLGregorianCalendar(GregorianCalendar.from(ZonedDateTime.now()));
+    var request = objectFactory.createRequestType();
+    var oaiPmhType = objectFactory.createOAIPMHtype();
+    oaiPmhType.setResponseDate(responseDate);
+    oaiPmhType.setRequest(request);
+
+    return objectFactory.createOAIPMH(oaiPmhType);
+  }
+}
